factor out nested RInstallation class
