+++ extensions/positron-r/src/util.ts
@@ -3,6 +3,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as vscode from 'vscode';
+import * as fs from 'fs';
 
 export function withActiveExtension(ext: vscode.Extension<any>, callback: () => void) {
 
@@ -28,5 +29,23 @@ export class PromiseHandles<T> {
 }
 
 export function delay(ms: number) {
-	return new Promise( resolve => setTimeout(resolve, ms) );
+	return new Promise(resolve => setTimeout(resolve, ms));
+}
+
+
+export function readLines(pth: string): Array<string> {
+	const bigString = fs.readFileSync(pth, 'utf8');
+	return bigString.split(/\r?\n/);
+}
+
+// extractValue('KEY=VALUE', 'KEY')      --> 'VALUE'
+// extractValue('KEY:VALUE', 'KEY', ':') --> 'VALUE'
+// extractValue('KEE:VALUE', 'KEY')      --> ''
+export function extractValue(str: string, key: string, delim: string = '='): string {
+	const re = `${key}${delim}(.*)`;
+	if (!str.startsWith(key)) {
+		return '';
+	}
+	const m = str.match(re);
+	return m?.[1] ?? '';
 }
